+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   concerns due to importing `dart:isolate`.
 * Remove `JsonDocumentTransformer` class. The `jsonDocument` top level is still
   available.
+* Remove `StreamChannelTransformer.typed`. Use `.cast` on the transformed
+  channel instead.
 
 ## 1.7.0